requirements.txt
faker
azure-kusto-ingest
azure-kusto-data